style: refactor LandingPage to use useAnimatedBackgroundGradient hook
