+++ src/index.tsx
@@ -9,6 +9,7 @@ import firebase from "firebase/app";
 
 import App from "./components/App";
 import { firebaseConfig } from "./firebaseConfig";
+import { useAnimatedBackgroundGradient } from "./components/common/useAnimatedBackgroundGradient";
 
 // import i18n (needs to be bundled ;))
 import "./i18n";
@@ -19,12 +20,35 @@ firebase.analytics().logEvent("screen_view", {
   screen_name: "index",
 } as any);
 
+function SuspenseFallback() {
+  useAnimatedBackgroundGradient();
+  return (
+    <div style={{
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "center",
+      justifyContent: "center",
+      height: "100vh"
+    }}>
+      <div style={{
+        width: 40,
+        height: 40,
+        border: "6px solid #e0e0e0",
+        borderTop: "6px solid #3f51b5",
+        borderRadius: "50%",
+        animation: "spin 1s linear infinite"
+      }} />
+      <style>{`@keyframes spin { 0% { transform: rotate(0deg); } 100% { transform: rotate(360deg); } }`}</style>
+    </div>
+  );
+}
+
 const container = document.getElementById("root");
 if (!container) throw new Error("Root container missing in index.html");
 const root = createRoot(container);
 root.render(
   <React.StrictMode>
-    <Suspense fallback={<div>Loading...</div>}>
+    <Suspense fallback={<SuspenseFallback />}>
       <App />
     </Suspense>
   </React.StrictMode>
